Those code now can calculate shipping fee using Giaohangnhanh API, but I didn't display it in jsp file yet

+++ src/main/java/vn/edu/hcmuaf/fit/sourcedoannoithat/dao/DetailDao.java
@@ -30,10 +30,10 @@ public class DetailDao {
                         rs.getString(4),
                         rs.getString(5),
                         rs.getString(6),
-                        rs.getDouble(7),
-                        rs.getDouble(8),
-                        rs.getDouble(9),
-                        rs.getDouble(10),
+                        rs.getInt(7),
+                        rs.getInt(8),
+                        rs.getInt(9),
+                        rs.getInt(10),
                         rs.getString(11),
                         rs.getString(12));
             }
